Move types to their own folder and files
