+++ src/components/Profile/FollowListModal.tsx
@@ -18,7 +18,7 @@ import {
 import { Link as Routerlink, useNavigate } from "react-router-dom";
 import PageLoader from "../../pages/PageLoader";
 import { useAppSelector } from "../../state/hooks";
-import { FollowableUser } from "../../state/slices/userSlice";
+import { FollowableUser } from "../../types/users";
 import FollowButton from "../Common/FollowButton";
 import FollowingButton from "../Common/FollowingButton";
 import SearchBar from "../Common/SearchBar";
