+++ src/components/Profile/ProfilePosts.tsx
@@ -1,7 +1,7 @@
 import { Divider, Stack } from "@mui/material";
 import { useEffect } from "react";
 import { useAppDispatch, useAppSelector } from "../../state/hooks";
-import { Post, setPosts } from "../../state/slices/postsSlice";
+import { setPosts } from "../../state/slices/postsSlice";
 import useAxios from "../../utilities/useAxios";
 import PostItem from "../Posts/PostItem";
 
@@ -23,7 +23,7 @@ const ProfilePosts = ({ visitedUserId }: ProfilePostsProps) => {
         },
         "profile/getUserPosts",
       );
-      dispatch(setPosts(result as Post[]));
+      dispatch(setPosts(result));
     };
     fetchPosts();
   }, [dispatch, visitedUserId]);
