+++ src/state/slices/postsSlice.ts
@@ -1,23 +1,5 @@
 import { PayloadAction, createSlice } from "@reduxjs/toolkit";
-
-export type Post = {
-  displayName: string;
-  followStatus: boolean;
-  imagePath?: string;
-  isLikedByCurrentUser: boolean;
-  isQuotePost?: boolean;
-  isRepost?: boolean;
-  numberOfLikes: number;
-  numberOfReplies: number;
-  numberOfReposts: number;
-  parentPostId?: number;
-  postId: number;
-  textContent: string;
-  timestamp: string;
-  userId: number;
-  username: string;
-  editedTimestamp: string;
-};
+import { Post } from "../../types/posts";
 
 type PostState = {
   posts: Post[];
