+++ src/state/slices/userSlice.ts
@@ -1,17 +1,6 @@
 import { PayloadAction, createSlice } from "@reduxjs/toolkit";
 
-export type AvatarUser = {
-  displayName: string;
-  imageUrl?: string;
-  userId: number;
-  username: string;
-};
-
-export type FollowableUser = AvatarUser & {
-  isFollowing: boolean;
-};
-
-export type CurrentUser = {
+type CurrentUser = {
   isLoading: boolean;
   userId?: number;
   username?: string;
