+++ src/types/users.ts
@@ -1,0 +1,10 @@
+export type AvatarUser = {
+  displayName: string;
+  imageUrl?: string;
+  userId: number;
+  username: string;
+};
+
+export type FollowableUser = AvatarUser & {
+  isFollowing: boolean;
+};
